updated/requirements.txt
pandas
snowflake-connector-python
streamlit